requirements.txt
Flask>=2.3
pillow
pytesseract
pandas
matplotlib
weasyprint
